+++ MatrixSDK/JSONModels/MXRoomCreateContent.m
@@ -33,6 +33,7 @@ static NSString* const kRoomCreateContentRoomTypeMSC1772JSONKey = @"org.matrix.m
 @property (nonatomic, strong, readwrite, nullable) MXRoomPredecessorInfo *roomPredecessorInfo;
 @property (nonatomic, copy, readwrite, nullable) NSString *roomVersion;
 @property (nonatomic, readwrite) BOOL isFederated;
+@property (nonatomic, readwrite, nonnull) MXVirtualRoomInfo *virtualRoomInfo;
 @property (nonatomic, readwrite, nullable) NSString *roomType;
 
 @end
@@ -50,7 +51,8 @@ static NSString* const kRoomCreateContentRoomTypeMSC1772JSONKey = @"org.matrix.m
         MXJSONModelSetString(roomCreateContent.creatorUserId, jsonDictionary[kRoomCreateContentUserIdJSONKey]);
         MXJSONModelSetMXJSONModel(roomCreateContent.roomPredecessorInfo, MXRoomPredecessorInfo, jsonDictionary[kRoomCreateContentPredecessorInfoJSONKey]);
         MXJSONModelSetString(roomCreateContent.roomVersion, jsonDictionary[kRoomCreateContentRoomVersionJSONKey]);
-        MXJSONModelSetBoolean(roomCreateContent.isFederated, jsonDictionary[kRoomCreateContentFederateJSONKey])
+        MXJSONModelSetBoolean(roomCreateContent.isFederated, jsonDictionary[kRoomCreateContentFederateJSONKey]);
+        MXJSONModelSetMXJSONModel(roomCreateContent.virtualRoomInfo, MXVirtualRoomInfo, jsonDictionary[kRoomIsVirtualJSONKey]);
         
         NSString *roomType;
         
@@ -82,6 +84,11 @@ static NSString* const kRoomCreateContentRoomTypeMSC1772JSONKey = @"org.matrix.m
     }
     
     jsonDictionary[kRoomCreateContentFederateJSONKey] = @(self.isFederated);
+    
+    if (self.virtualRoomInfo)
+    {
+        jsonDictionary[kRoomIsVirtualJSONKey] = self.virtualRoomInfo;
+    }
     
     if (self.roomType)
     {