+++ MatrixSDK/MXSession.h
@@ -38,6 +38,7 @@
 #import "MXIdentityService.h"
 #import "MX3PidAddManager.h"
 #import "MXMembershipTransitionState.h"
+#import "MXRoomAccountDataUpdating.h"
 
 /**
  `MXSessionState` represents the states in the life cycle of a MXSession instance.
@@ -213,6 +214,13 @@ FOUNDATION_EXPORT NSString *const kMXSessionIgnoredUsersDidChangeNotification;
  The notification object is the concerned session (MXSession instance).
  */
 FOUNDATION_EXPORT NSString *const kMXSessionDirectRoomsDidChangeNotification;
+
+/**
+ Posted when the virtual rooms are updated, either from the store or from the homeserver.
+ 
+ The notification object is the concerned session (MXSession instance).
+ */
+FOUNDATION_EXPORT NSString *const kMXSessionVirtualRoomsDidChangeNotification;
 
 /**
  Posted when the matrix account data are updated from homeserver.
@@ -1020,6 +1028,12 @@ typedef void (^MXOnBackgroundSyncFail)(NSError *error);
  */
 @property id<MXRoomSummaryUpdating> roomSummaryUpdateDelegate;
 
+/**
+ Delegate for updating room account data.
+ By default, it is the one returned by [MXRoomAccountDataUpdater roomAccountDataUpdaterForSession:].
+ */
+@property id<MXRoomAccountDataUpdating> roomAccountDataUpdateDelegate;
+
 #pragma mark - The user's groups
 /**
  Get the MXGroup instance of a user's group.
@@ -1502,5 +1516,24 @@ typedef void (^MXOnBackgroundSyncFail)(NSError *error);
  @return the current publicised groups for the provided user.
  */
 - (NSArray<NSString *> *)publicisedGroupsForUser:(NSString*)userId;
+
+#pragma mark - Virtual Rooms
+
+/**
+ Cache the virtual room of a native room.
+ 
+ @param virtualRoomId nirtual room identifier
+ @param nativeRoomId native room identifier.
+ */
+- (void)setVirtualRoom:(NSString *)virtualRoomId
+         forNativeRoom:(NSString *)nativeRoomId;
+
+/**
+ Get virtual room identifier for a given native room identifier.
+ 
+ @param nativeRoomId native room identifier to look for the virtual room.
+ @return the virtual room identifier for the given native room. May be nil.
+ */
+- (NSString *)virtualRoomOf:(NSString *)nativeRoomId;
 
 @end